app/test/langgraph-wechat-assistant/requirements.txt
fastapi
uvicorn
redis
openai
langchain
langgraph
transformers
torch
python-dotenv
requests
python-dotenv